upate 2.2
checkCells has been updated to checkCells3 with the new method in place.

test_checkCells is a program that checks a random file and a random frame in that file, changes the coordinates a little, then tries to find the closest frame to that. User can go in test_checkCells.f90 and fiddle the coordinates as they want.

To execute, do "make -f make_test_checkCells" and execute the a.out. output is in the progressfile as specified in f1_parameters.f90.

 	modified:   checkCells3.f90
 	make_test_checkCells
 	test_checkCells.f90